+++ radicle-surf/Cargo.toml
@@ -15,6 +15,10 @@ include = [
     "data/git-platinum.tgz",
 ]
 
+[lib]
+test = false
+doctest = false
+
 [features]
 serialize = ["serde"]
 # NOTE: testing `test_submodule_failure` on GH actions
