+++ radicle-surf/src/vcs/git.rs
@@ -69,14 +69,14 @@ pub use radicle_git_ext::Oid;
 
 /// Provides ways of selecting a particular reference/revision.
 mod reference;
-pub use reference::{Ref, Rev};
+pub use reference::{ParseError, Ref, Rev};
 
 mod repo;
 pub use repo::{History, Repository, RepositoryRef};
 
 pub mod error;
 
-mod ext;
+pub mod ext;
 
 /// Provides the data for talking about branches.
 pub mod branch;
@@ -1179,593 +1179,5 @@ impl<'a> Browser<'a> {
                 size: blob.size(),
             },
         ))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[cfg(not(feature = "gh-actions"))]
-    #[test]
-    // An issue with submodules, see: https://github.com/radicle-dev/radicle-surf/issues/54
-    fn test_submodule_failure() {
-        let repo = Repository::new("..").unwrap();
-        let browser = Browser::new(&repo, Branch::local("main")).unwrap();
-
-        browser.get_directory().unwrap();
-    }
-
-    #[cfg(test)]
-    mod namespace {
-        use super::*;
-        use pretty_assertions::{assert_eq, assert_ne};
-
-        #[test]
-        fn switch_to_banana() -> Result<(), Error> {
-            let repo = Repository::new("./data/git-platinum")?;
-            let mut browser = Browser::new_with_namespace(
-                &repo,
-                &Namespace::try_from("golden")?,
-                Branch::local("master"),
-            )?;
-            let history = browser.history.clone();
-
-            browser.branch(Branch::local("banana"))?;
-
-            assert_ne!(history, browser.history);
-
-            Ok(())
-        }
-
-        #[test]
-        fn me_namespace() -> Result<(), Error> {
-            let repo = Repository::new("./data/git-platinum")?;
-            let browser = Browser::new(&repo, Branch::local("master"))?;
-            let history = browser.history.clone();
-
-            assert_eq!(browser.which_namespace(), Ok(None));
-
-            let browser = browser
-                .switch_namespace(&Namespace::try_from("me")?, Branch::local("feature/#1194"))?;
-
-            assert_eq!(
-                browser.which_namespace(),
-                Ok(Some(Namespace::try_from("me")?))
-            );
-            assert_eq!(history, browser.history);
-
-            let expected_branches: Vec<Branch> = vec![Branch::local("feature/#1194")];
-            let mut branches = browser.list_branches(RefScope::Local)?;
-            branches.sort();
-
-            assert_eq!(expected_branches, branches);
-
-            let expected_branches: Vec<Branch> = vec![Branch::remote("feature/#1194", "fein")];
-            let mut branches = browser.list_branches(RefScope::Remote {
-                name: Some("fein".to_string()),
-            })?;
-            branches.sort();
-
-            assert_eq!(expected_branches, branches);
-
-            Ok(())
-        }
-
-        #[test]
-        fn golden_namespace() -> Result<(), Error> {
-            let repo = Repository::new("./data/git-platinum")?;
-            let browser = Browser::new(&repo, Branch::local("master"))?;
-            let history = browser.history.clone();
-
-            assert_eq!(browser.which_namespace(), Ok(None));
-
-            let golden_browser = browser
-                .switch_namespace(&Namespace::try_from("golden")?, Branch::local("master"))?;
-
-            assert_eq!(
-                golden_browser.which_namespace(),
-                Ok(Some(Namespace::try_from("golden")?))
-            );
-            assert_eq!(history, golden_browser.history);
-
-            let expected_branches: Vec<Branch> =
-                vec![Branch::local("banana"), Branch::local("master")];
-            let mut branches = golden_browser.list_branches(RefScope::Local)?;
-            branches.sort();
-
-            assert_eq!(expected_branches, branches);
-
-            let expected_branches: Vec<Branch> = vec![
-                Branch::remote("fakie/bigspin", "kickflip"),
-                Branch::remote("heelflip", "kickflip"),
-                Branch::remote("v0.1.0", "kickflip"),
-            ];
-            let mut branches = golden_browser.list_branches(RefScope::Remote {
-                name: Some("kickflip".to_string()),
-            })?;
-            branches.sort();
-
-            assert_eq!(expected_branches, branches);
-
-            Ok(())
-        }
-
-        #[test]
-        fn silver_namespace() -> Result<(), Error> {
-            let repo = Repository::new("./data/git-platinum")?;
-            let browser = Browser::new(&repo, Branch::local("master"))?;
-            let history = browser.history.clone();
-
-            assert_eq!(browser.which_namespace(), Ok(None));
-
-            let silver_browser = browser.switch_namespace(
-                &Namespace::try_from("golden/silver")?,
-                Branch::local("master"),
-            )?;
-
-            assert_eq!(
-                silver_browser.which_namespace(),
-                Ok(Some(Namespace::try_from("golden/silver")?))
-            );
-            assert_ne!(history, silver_browser.history);
-
-            let expected_branches: Vec<Branch> = vec![Branch::local("master")];
-            let mut branches = silver_browser.list_branches(RefScope::All)?;
-            branches.sort();
-
-            assert_eq!(expected_branches, branches);
-
-            Ok(())
-        }
-    }
-
-    #[cfg(test)]
-    mod rev {
-        use super::{Branch, Browser, Error, Oid, Repository, TagName};
-        use std::str::FromStr;
-
-        // **FIXME**: This seems to break occasionally on
-        // buildkite. For some reason the commit
-        // 3873745c8f6ffb45c990eb23b491d4b4b6182f95, which is on master
-        // (currently HEAD), is not found. It seems to load the history
-        // with d6880352fc7fda8f521ae9b7357668b17bb5bad5 as the HEAD.
-        //
-        // To temporarily fix this, we need to select "New Build" from the build kite
-        // build page that's failing.
-        // * Under "Message" put whatever you want.
-        // * Under "Branch" put in the branch you're working on.
-        // * Expand "Options" and select "clean checkout".
-        #[test]
-        fn _master() -> Result<(), Error> {
-            let repo = Repository::new("./data/git-platinum")?;
-            let mut browser = Browser::new(&repo, Branch::local("master"))?;
-            browser.rev(Branch::remote("master", "origin"))?;
-
-            let commit1 = Oid::from_str("3873745c8f6ffb45c990eb23b491d4b4b6182f95")?;
-            assert!(
-                browser
-                    .history
-                    .find(|commit| if commit.id == commit1 {
-                        Some(commit.clone())
-                    } else {
-                        None
-                    })
-                    .is_some(),
-                "commit_id={}, history =\n{:#?}",
-                commit1,
-                browser.history
-            );
-
-            let commit2 = Oid::from_str("d6880352fc7fda8f521ae9b7357668b17bb5bad5")?;
-            assert!(
-                browser
-                    .history
-                    .find(|commit| if commit.id == commit2 {
-                        Some(commit.clone())
-                    } else {
-                        None
-                    })
-                    .is_some(),
-                "commit_id={}, history =\n{:#?}",
-                commit2,
-                browser.history
-            );
-
-            Ok(())
-        }
-
-        #[test]
-        fn commit() -> Result<(), Error> {
-            let repo = Repository::new("./data/git-platinum")?;
-            let mut browser = Browser::new(&repo, Branch::local("master"))?;
-            browser.rev(Oid::from_str("3873745c8f6ffb45c990eb23b491d4b4b6182f95")?)?;
-
-            let commit1 = Oid::from_str("3873745c8f6ffb45c990eb23b491d4b4b6182f95")?;
-            assert!(browser
-                .history
-                .find(|commit| if commit.id == commit1 {
-                    Some(commit.clone())
-                } else {
-                    None
-                })
-                .is_some());
-
-            Ok(())
-        }
-
-        #[test]
-        fn commit_parents() -> Result<(), Error> {
-            let repo = Repository::new("./data/git-platinum")?;
-            let mut browser = Browser::new(&repo, Branch::local("master"))?;
-            browser.rev(Oid::from_str("3873745c8f6ffb45c990eb23b491d4b4b6182f95")?)?;
-            let commit = browser.history.first();
-
-            assert_eq!(
-                commit.parents,
-                vec![Oid::from_str("d6880352fc7fda8f521ae9b7357668b17bb5bad5")?]
-            );
-
-            Ok(())
-        }
-
-        #[test]
-        fn commit_short() -> Result<(), Error> {
-            let repo = Repository::new("./data/git-platinum")?;
-            let mut browser = Browser::new(&repo, Branch::local("master"))?;
-            browser.rev(browser.oid("3873745c8")?)?;
-
-            let commit1 = Oid::from_str("3873745c8f6ffb45c990eb23b491d4b4b6182f95")?;
-            assert!(browser
-                .history
-                .find(|commit| if commit.id == commit1 {
-                    Some(commit.clone())
-                } else {
-                    None
-                })
-                .is_some());
-
-            Ok(())
-        }
-
-        #[test]
-        fn tag() -> Result<(), Error> {
-            let repo = Repository::new("./data/git-platinum")?;
-            let mut browser = Browser::new(&repo, Branch::local("master"))?;
-            browser.rev(TagName::new("v0.2.0"))?;
-
-            let commit1 = Oid::from_str("2429f097664f9af0c5b7b389ab998b2199ffa977")?;
-            assert_eq!(browser.history.first().id, commit1);
-
-            Ok(())
-        }
-    }
-
-    #[cfg(test)]
-    mod last_commit {
-        use crate::{
-            file_system::{unsound, Path},
-            vcs::git::{Branch, Browser, Oid, Repository},
-        };
-        use std::str::FromStr;
-
-        #[test]
-        fn readme_missing_and_memory() {
-            let repo = Repository::new("./data/git-platinum")
-                .expect("Could not retrieve ./data/git-platinum as git repository");
-            let mut browser =
-                Browser::new(&repo, Branch::local("master")).expect("Could not initialise Browser");
-
-            // Set the browser history to the initial commit
-            let commit = Oid::from_str("d3464e33d75c75c99bfb90fa2e9d16efc0b7d0e3")
-                .expect("Failed to parse SHA");
-            browser.commit(commit).unwrap();
-
-            let head_commit = browser.get().0.first().clone();
-
-            // memory.rs is commited later so it should not exist here.
-            let memory_last_commit = browser
-                .last_commit(Path::with_root(&[
-                    unsound::label::new("src"),
-                    unsound::label::new("memory.rs"),
-                ]))
-                .expect("Failed to get last commit")
-                .map(|commit| commit.id);
-
-            assert_eq!(memory_last_commit, None);
-
-            // README.md exists in this commit.
-            let readme_last_commit = browser
-                .last_commit(Path::with_root(&[unsound::label::new("README.md")]))
-                .expect("Failed to get last commit")
-                .map(|commit| commit.id);
-
-            assert_eq!(readme_last_commit, Some(head_commit.id));
-        }
-
-        #[test]
-        fn folder_svelte() {
-            let repo = Repository::new("./data/git-platinum")
-                .expect("Could not retrieve ./data/git-platinum as git repository");
-            let mut browser =
-                Browser::new(&repo, Branch::local("master")).expect("Could not initialise Browser");
-
-            // Check that last commit is the actual last commit even if head commit differs.
-            let commit = Oid::from_str("19bec071db6474af89c866a1bd0e4b1ff76e2b97")
-                .expect("Could not parse SHA");
-            browser.commit(commit).unwrap();
-
-            let expected_commit_id =
-                Oid::from_str("f3a089488f4cfd1a240a9c01b3fcc4c34a4e97b2").unwrap();
-
-            let folder_svelte = browser
-                .last_commit(unsound::path::new("~/examples/Folder.svelte"))
-                .expect("Failed to get last commit")
-                .map(|commit| commit.id);
-
-            assert_eq!(folder_svelte, Some(expected_commit_id));
-        }
-
-        #[test]
-        fn nest_directory() {
-            let repo = Repository::new("./data/git-platinum")
-                .expect("Could not retrieve ./data/git-platinum as git repository");
-            let mut browser =
-                Browser::new(&repo, Branch::local("master")).expect("Could not initialise Browser");
-
-            // Check that last commit is the actual last commit even if head commit differs.
-            let commit = Oid::from_str("19bec071db6474af89c866a1bd0e4b1ff76e2b97")
-                .expect("Failed to parse SHA");
-            browser.commit(commit).unwrap();
-
-            let expected_commit_id =
-                Oid::from_str("2429f097664f9af0c5b7b389ab998b2199ffa977").unwrap();
-
-            let nested_directory_tree_commit_id = browser
-                .last_commit(unsound::path::new(
-                    "~/this/is/a/really/deeply/nested/directory/tree",
-                ))
-                .expect("Failed to get last commit")
-                .map(|commit| commit.id);
-
-            assert_eq!(nested_directory_tree_commit_id, Some(expected_commit_id));
-        }
-
-        #[test]
-        #[cfg(not(windows))]
-        fn can_get_last_commit_for_special_filenames() {
-            let repo = Repository::new("./data/git-platinum")
-                .expect("Could not retrieve ./data/git-platinum as git repository");
-            let mut browser =
-                Browser::new(&repo, Branch::local("master")).expect("Could not initialise Browser");
-
-            // Check that last commit is the actual last commit even if head commit differs.
-            let commit = Oid::from_str("a0dd9122d33dff2a35f564d564db127152c88e02")
-                .expect("Failed to parse SHA");
-            browser.commit(commit).unwrap();
-
-            let expected_commit_id =
-                Oid::from_str("a0dd9122d33dff2a35f564d564db127152c88e02").unwrap();
-
-            let backslash_commit_id = browser
-                .last_commit(unsound::path::new("~/special/faux\\path"))
-                .expect("Failed to get last commit")
-                .map(|commit| commit.id);
-            assert_eq!(backslash_commit_id, Some(expected_commit_id));
-
-            let ogre_commit_id = browser
-                .last_commit(unsound::path::new("~/special/👹👹👹"))
-                .expect("Failed to get last commit")
-                .map(|commit| commit.id);
-            assert_eq!(ogre_commit_id, Some(expected_commit_id));
-        }
-
-        #[test]
-        fn root() {
-            let repo = Repository::new("./data/git-platinum")
-                .expect("Could not retrieve ./data/git-platinum as git repository");
-            let browser =
-                Browser::new(&repo, Branch::local("master")).expect("Could not initialise Browser");
-
-            let root_last_commit_id = browser
-                .last_commit(Path::root())
-                .expect("Failed to get last commit")
-                .map(|commit| commit.id);
-
-            assert_eq!(root_last_commit_id, Some(browser.get().first().id));
-        }
-    }
-
-    #[cfg(test)]
-    mod diff {
-        use crate::{diff::*, vcs::git::*};
-        use std::str::FromStr;
-
-        #[test]
-        fn test_initial_diff() -> Result<(), Error> {
-            use file_system::*;
-            use pretty_assertions::assert_eq;
-
-            let oid = Oid::from_str("d3464e33d75c75c99bfb90fa2e9d16efc0b7d0e3")?;
-            let repo = Repository::new("./data/git-platinum")?;
-            let commit = repo.0.find_commit(oid.into()).unwrap();
-
-            assert!(commit.parents().count() == 0);
-            assert!(commit.parent(0).is_err());
-
-            let bro = Browser::new(&repo, Branch::local("master"))?;
-            let diff = bro.initial_diff(oid)?;
-
-            let expected_diff = Diff {
-                created: vec![CreateFile {
-                    path: Path::with_root(&[unsound::label::new("README.md")]),
-                    diff: FileDiff::Plain {
-                        hunks: vec![Hunk {
-                            header: Line(b"@@ -0,0 +1 @@\n".to_vec()),
-                            lines: vec![
-                                LineDiff::addition(b"This repository is a data source for the Upstream front-end tests.\n".to_vec(), 1),
-                            ]
-                        }].into()
-                    },
-                }],
-                deleted: vec![],
-                moved: vec![],
-                copied: vec![],
-                modified: vec![],
-            };
-            assert_eq!(expected_diff, diff);
-
-            Ok(())
-        }
-
-        #[test]
-        fn test_diff() -> Result<(), Error> {
-            use file_system::*;
-            use pretty_assertions::assert_eq;
-
-            let repo = Repository::new("./data/git-platinum")?;
-            let commit = repo
-                .0
-                .find_commit(git2::Oid::from_str(
-                    "80bacafba303bf0cdf6142921f430ff265f25095",
-                )?)
-                .unwrap();
-            let parent = commit.parent(0)?;
-
-            let bro = Browser::new(&repo, Branch::local("master"))?;
-
-            let diff = bro.diff(parent.id().into(), commit.id().into())?;
-
-            let expected_diff = Diff {
-                created: vec![],
-                deleted: vec![],
-                moved: vec![],
-                copied: vec![],
-                modified: vec![ModifiedFile {
-                    path: Path::with_root(&[unsound::label::new("README.md")]),
-                    diff: FileDiff::Plain {
-                        hunks: vec![Hunk {
-                            header: Line(b"@@ -1 +1,2 @@\n".to_vec()),
-                            lines: vec![
-                                LineDiff::deletion(b"This repository is a data source for the Upstream front-end tests.\n".to_vec(), 1),
-                                LineDiff::addition(b"This repository is a data source for the Upstream front-end tests and the\n".to_vec(), 1),
-                                LineDiff::addition(b"[`radicle-surf`](https://github.com/radicle-dev/git-platinum) unit tests.\n".to_vec(), 2),
-                            ]
-                        }].into()
-                    },
-                    eof: None,
-                }]
-            };
-            assert_eq!(expected_diff, diff);
-
-            Ok(())
-        }
-
-        #[cfg(feature = "serialize")]
-        #[test]
-        fn test_diff_serde() -> Result<(), Error> {
-            use file_system::*;
-
-            let diff = Diff {
-                created: vec![CreateFile{path: unsound::path::new("LICENSE"), diff: FileDiff::Plain { hunks: Hunks::default() }}],
-                deleted: vec![],
-                moved: vec![
-                    MoveFile {
-                        old_path: unsound::path::new("CONTRIBUTING"),
-                        new_path: unsound::path::new("CONTRIBUTING.md")
-                    }
-                ],
-                copied: vec![],
-                modified: vec![ModifiedFile {
-                    path: Path::with_root(&[unsound::label::new("README.md")]),
-                    diff: FileDiff::Plain {
-                        hunks: vec![Hunk {
-                            header: Line(b"@@ -1 +1,2 @@\n".to_vec()),
-                            lines: vec![
-                                LineDiff::deletion(b"This repository is a data source for the Upstream front-end tests.\n".to_vec(), 1),
-                                LineDiff::addition(b"This repository is a data source for the Upstream front-end tests and the\n".to_vec(), 1),
-                                LineDiff::addition(b"[`radicle-surf`](https://github.com/radicle-dev/git-platinum) unit tests.\n".to_vec(), 2),
-                                LineDiff::context(b"\n".to_vec(), 3, 4),
-                            ]
-                        }].into()
-                    },
-                    eof: None,
-                }]
-            };
-
-            let eof: Option<u8> = None;
-            let json = serde_json::json!({
-                "created": [{"path": "LICENSE", "diff": {
-                        "type": "plain",
-                        "hunks": []
-                    },
-                }],
-                "deleted": [],
-                "moved": [{ "oldPath": "CONTRIBUTING", "newPath": "CONTRIBUTING.md" }],
-                "copied": [],
-                "modified": [{
-                    "path": "README.md",
-                    "diff": {
-                        "type": "plain",
-                        "hunks": [{
-                            "header": "@@ -1 +1,2 @@\n",
-                            "lines": [
-                                { "lineNum": 1,
-                                  "line": "This repository is a data source for the Upstream front-end tests.\n",
-                                  "type": "deletion"
-                                },
-                                { "lineNum": 1,
-                                  "line": "This repository is a data source for the Upstream front-end tests and the\n",
-                                  "type": "addition"
-                                },
-                                { "lineNum": 2,
-                                  "line": "[`radicle-surf`](https://github.com/radicle-dev/git-platinum) unit tests.\n",
-                                  "type": "addition"
-                                },
-                                { "lineNumOld": 3, "lineNumNew": 4,
-                                  "line": "\n",
-                                  "type": "context"
-                                }
-                            ]
-                        }]
-                    },
-                    "eof" : eof,
-                }]
-            });
-            assert_eq!(serde_json::to_value(&diff).unwrap(), json);
-
-            Ok(())
-        }
-    }
-
-    #[cfg(test)]
-    mod threading {
-        use crate::vcs::git::*;
-        use pretty_assertions::assert_eq;
-        use std::sync::{Mutex, MutexGuard};
-
-        #[test]
-        fn basic_test() -> Result<(), Error> {
-            let shared_repo = Mutex::new(Repository::new("./data/git-platinum")?);
-            let locked_repo: MutexGuard<Repository> = shared_repo.lock().unwrap();
-            let bro = Browser::new(&*locked_repo, Branch::local("master"))?;
-            let mut branches = bro.list_branches(RefScope::All)?;
-            branches.sort();
-
-            assert_eq!(
-                branches,
-                vec![
-                    Branch::remote("HEAD", "origin"),
-                    Branch::remote("dev", "origin"),
-                    Branch::local("dev"),
-                    Branch::remote("master", "origin"),
-                    Branch::local("master"),
-                    Branch::remote("orange/pineapple", "banana"),
-                    Branch::remote("pineapple", "banana"),
-                ]
-            );
-
-            Ok(())
-        }
     }
 }
