+++ radicle-surf/t/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "radicle-surf-test"
+version = "0.1.0"
+edition = "2021"
+license = "GPL-3.0-or-later"
+
+publish = false
+
+[lib]
+test = true
+
+[features]
+serialize = []
+test = []
+
+[dev-dependencies]
+nonempty = "0.5"
+pretty_assertions = "1.3.0"
+proptest = "0.9"
+serde_json = "1"
+
+[dev-dependencies.git2]
+version = "0.15.0"
+default-features = false
+features = ["vendored-libgit2"]
+
+[dev-dependencies.radicle-git-ext]
+path = "../../radicle-git-ext"
+
+[dev-dependencies.radicle-surf]
+path = ".."
+features = ["serialize"]
+
+[dev-dependencies.test-helpers]
+path = "../../test/test-helpers"
