+++ radicle-surf/t/src/lib.rs
@@ -1,0 +1,14 @@
+// Copyright © 2022 The Radicle Git Contributors
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#[cfg(test)]
+mod git;
+
+#[cfg(test)]
+mod diff;
+
+#[cfg(test)]
+mod file_system;
+
+#[cfg(test)]
+mod tree;
